behavioral_auth/requirements.txt
# requirements.txt — Python dependencies
# Install with: pip install -r requirements.txt

Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9          # PostgreSQL adapter
python-dotenv==1.0.1            # For loading .env variables
